Filled out English setup instructions
I translated the rest of the Russian setup instructions into English.

Unfortunately, since I don't currently have access to an Android device
to take screenshots on, several of the images are still in Russian. Once
Enlish screenshots can be taken, this should be very easy to update
(just add the new images to the folder, and update any link ending in
"ru" to end in "en").